+++ src/system/unix/unix_system.c
@@ -178,7 +178,7 @@ void sys_show_console(int viewlog, int b)
 }
 
 void *sys_load_dll(const char *name, char *fqpath, 
-	int (**entrypoint)(intptr_t, ...), int (*systemcalls)(intptr_t, ...))
+	intptr_t (**entrypoint)(intptr_t, ...), intptr_t (*systemcalls)(intptr_t, ...))
 {
 	void *libhandle;
 	void (*dllentry)(intptr_t (*syscallptr)(intptr_t, ...));
